requirements.txt
# Python dependencies for SecurityScanner
httpx>=0.24.0
pyyaml>=6.0
requests>=2.31.0
urllib3>=2.0.0
python-owasp-zap-v2.4>=0.0.21
